Added optional tomcat ROOT redirect script

+++ bin/scripts/centos7/install.sh
@@ -147,17 +147,36 @@ case "$clientchoice" in
         echo '... skipping Keyterms client installation.'
         ;;
     *) echo '... deploying Keyterms client to Tomcat ...'
-        mv $CATALINA_HOME/webapps/ROOT $CATALINA_HOME/webapps/ROOT_BAK
-        CLIENT_TOMCAT_DIR=$CATALINA_HOME/webapps/ROOT
+        CLIENT_TOMCAT_DIR=$CATALINA_HOME/webapps/keyterms
         mkdir -p $CLIENT_TOMCAT_DIR
         cp -R $CLIENT_DIR/public/keyterms/* $CLIENT_TOMCAT_DIR
         chown -R $TOMCAT_USER:$APP_GROUP $CLIENT_TOMCAT_DIR
 
-        echo "... KeyTerms client has been placed into $CLIENT_TOMCAT_DIR"
+        echo '... copying in config file ...'
         cp $CONF_DIR/client-config.js $CLIENT_TOMCAT_DIR/config.js
         serverURL="$SV_PROTOCOL://$hostname:$SV_PORT/"
         sed -i -e "s|myServerLocation|${serverURL}|g" $CLIENT_TOMCAT_DIR/config.js
-        echo "USER ATTENTION REQUIRED: If you are having problems with the client under tomcat, please verify the server location setting in $CATALINA_HOME/webapps/ROOT/keyterms/config.js"
+        echo "... KeyTerms client deployed to $CLIENT_TOMCAT_DIR"
+        echo "USER ATTENTION REQUIRED: If you are having problems with the client under tomcat, please verify the server location setting in $CLIENT_TOMCAT_DIR/config.js"
+
+        echo ' '; read -p 'Would you like to install a script in the ROOT webapp to redirect to the KeyTerms client? (y|N) '
+        y|Y)
+            echo '... backing up ROOT webapp ...'
+            ROOT_DIR=$CATALINA_HOME/webapps/ROOT
+            if ! [ -d $CATALINA_HOME/webapps/ROOT_BAK ]; then
+                mv $CATALINA_HOME/webapps/ROOT $CATALINA_HOME/webapps/ROOT_BAK
+            fi
+            echo '... configuring ROOT for KeyTerms redirect ...'
+            cp $CONF_DIR/root-default-web.xml $ROOT_DIR/WEB-INF/web.xml
+            if [ -e $ROOT_DIR/index.html ]; then
+                rm $ROOT_DIR/index.html
+            fi
+            cp $CONF_DIR/root-redirect-index.jsp $ROOT_DIR/index.jsp
+            echo '... done. ROOT will now redirect to KeyTerms'
+            ;;
+        *)
+            echo '... not installing redirect. ROOT webapp will remain accessible.'
+            ;;
         ;;
 esac
